Refactor folders to be feature oriented
