+++ packages/vtex-io-logger/src/core/injection/enhancedLoggerInjectionMiddlewareFactory.ts
